Fix printf int specifier but size_t warnings
Fix is C89-friendly.  Could cause problems if there
are more tests than an unsigned long can handle, but
I seriously doubt someone would write that many tests.

+++ include/fct.h
@@ -2902,10 +2902,10 @@ fct_standard_logger__on_fctx_end(
     num_tests = fctkern__tst_cnt(e->kern);
     num_passed = fctkern__tst_cnt_passed(e->kern);
     printf(
-        "%s (%d/%d tests",
+        "%s (%lu/%lu tests",
         (is_success) ? "PASSED" : "FAILED",
-        num_passed,
-        num_tests
+        (unsigned long) num_passed,
+        (unsigned long) num_tests
     );
     elasped_time = fct_timer__duration(&(logger->timer));
     if ( elasped_time > 0.0000001 )
@@ -3018,10 +3018,11 @@ fct_junit_logger__on_test_suite_end(
     FCT_SWITCH_STDERR_TO_STDERR();
 
     /* opening testsuite tag */
-    printf("\t<testsuite errors=\"%d\" failures=\"0\" tests=\"%d\" "
+    printf("\t<testsuite errors=\"%lu\" failures=\"0\" tests=\"%lu\" "
            "name=\"%s\" time=\"%.4f\">\n",
-           fct_ts__tst_cnt(ts) - fct_ts__tst_cnt_passed(ts),
-           fct_ts__tst_cnt(ts),
+           (unsigned long)   fct_ts__tst_cnt(ts)
+                           - fct_ts__tst_cnt_passed(ts),
+           (unsigned long) fct_ts__tst_cnt(ts),
            fct_ts__name(ts),
            elasped_time);
